% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assessment_survial_model}
\alias{assessment_survial_model}
\title{assessment_survial_model}
\usage{
assessment_survial_model(
  tdata,
  genes,
  coefs,
  cut_method = "best",
  out_name = "Survial",
  label.x = 4,
  label.y = 0.95,
  km_cols = c("#cd0000", "#118aff"),
  roc_cols = c("#1C79B7", "#F38329", "#2DA248"),
  time_point = c(1, 3, 5)
)
}
\description{
assessment_survial_model
}
\author{
Jiahao Wang
}
