% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{bin2isi}
\alias{bin2isi}
\title{Convert a binary array of spike indicators to ISIs}
\usage{
bin2isi(Y, dt, i = NULL)
}
\arguments{
\item{Y}{A \verb{q x n x r} or \verb{n x r} array of 0/1 spike trains
where \code{q} is the number of cells,
\code{n} is the number of time bins, and \code{r} is the number of trials.}

\item{dt}{Scalar of length of the bins.}

\item{i}{Index of the neuron for which ISI will be obtained. Only need
to be provided if Y is 3D.}
}
\value{
A vector of ISIs.
}
\description{
Convert a binary array of spike indicators to ISIs
}
