% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{bin2num}
\alias{bin2num}
\title{Convert binary spike data to spike time data}
\usage{
bin2num(spikes, dt)
}
\arguments{
\item{spikes}{A vector of 0s and 1s with 1 indicating a spike, or a matrix / data frame
of \code{r} rows of such vectors, with \code{r} being the number of trials.}

\item{dt}{A scalar of length of the time bins.}
}
\value{
A vector or list of spike times.
}
\description{
Convert binary spike data to spike time data
}
