% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_percentile_point}
\alias{calculate_percentile_point}
\title{calculate_percentile_point}
\usage{
calculate_percentile_point(A, B, percentile)
}
\description{
calculate_percentile_point
}
\author{
Jiahao Wang
}
