% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{cluster_start_with_one}
\alias{cluster_start_with_one}
\title{cluster_start_with_one}
\usage{
cluster_start_with_one(obj, cluster_name = "seurat_clusters")

cluster_start_with_one(obj, cluster_name = "seurat_clusters")
}
\description{
cluster_start_with_one

cluster_start_with_one
}
\author{
Jiahao Wang
}
