% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_fastR.R
\name{combo_fastR}
\alias{combo_fastR}
\title{Combine sheets from FAST - Intervention, Initiative, CODB}
\usage{
combo_fastR(filepath)
}
\arguments{
\item{filepath}{full file path for the COP19 FAST}
}
\description{
Combine sheets from FAST - Intervention, Initiative, CODB
}
\examples{
\dontrun{
#FAST file path
  path <- "../Downloads/FAST Malawi Consolidated Zero Draft 02082019.xlsx"
#read in FAST
  df_fast <- combo_fastR(path) }
}
