% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Rnll.R
\name{compute_Rnll}
\alias{compute_Rnll}
\title{Compute the negative log-likelihood of the factor model in R}
\usage{
compute_Rnll(data, param_list)
}
\arguments{
\item{data}{A list of a 3d data array \code{Y}, penalization parameter \code{lam}, time bin size \code{dt}, and num of factors \code{n_factor}.}

\item{param_list}{A list of parameters including vector \code{log_k}, vector \code{log_a}, 3d array \code{x} (latent paths),
and vector \code{Lt} (unnormalized lower tri elements of loading matrix)}
}
\value{
Scalar of negative log-likelihood
}
\description{
Compute the negative log-likelihood of the factor model in R
}
