% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R, R/utils.R
\name{convertEnrichT}
\alias{convertEnrichT}
\title{convertEnrichT}
\usage{
convertEnrichT(type, enrichT, species = "human")

convertEnrichT(type, enrichT, species = "human")
}
\description{
convertEnrichT

convertEnrichT
}
\author{
Jiahao Wang
}
