% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R, R/utils.R
\name{convertId}
\alias{convertId}
\title{convertId}
\usage{
convertId(
  genes = NULL,
  from,
  to,
  OrgDb = "org.Hs.eg.db",
  drop = TRUE,
  remove = TRUE
)

convertId(
  genes = NULL,
  from,
  to,
  OrgDb = "org.Hs.eg.db",
  drop = TRUE,
  remove = TRUE
)
}
\description{
convertId

convertId
}
\author{
Jiahao Wang
}
