% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cor2real}
\alias{cor2real}
\title{Transform the correlation coefficient to a real number}
\usage{
cor2real(x)
}
\arguments{
\item{x}{Correlation coefficient that lies in (-1, 1).}
}
\value{
Transformed x, a real number.
}
\description{
Transform the correlation coefficient to a real number
}
\details{
\code{x} is first transformed to the scale of (0, 1), and then is applied to a logit function.
}
