% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{creatSeuratObj}
\alias{creatSeuratObj}
\title{creatSeuratObj}
\usage{
creatSeuratObj(countFile, metaFile)

creatSeuratObj(countFile, metaFile)
}
\description{
creatSeuratObj

creatSeuratObj
}
\author{
Jiahao Wang
}
