% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DimPlot2}
\alias{DimPlot2}
\title{DimPlot2}
\usage{
DimPlot2(
  obj,
  group.by,
  legend.title = NULL,
  title = "All",
  cols = col.cluster2.1,
  reduction = "umap",
  cex = 20,
  size = 0.3,
  label = TRUE,
  label.fill = "auto",
  label.cex = 3,
  mini.axis = FALSE,
  legend = TRUE,
  pos.manual = NULL,
  ...
)
}
\description{
DimPlot2
}
\author{
Jiahao Wang
}
