% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_TCGA}
\alias{download_TCGA}
\title{download_TCGA}
\usage{
download_TCGA(
  project,
  data_type,
  save.filename = NULL,
  directory = "./",
  dedup = "mean"
)
}
\description{
download_TCGA
}
\author{
Jiahao Wang
}
