% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ou.R
\name{extract_ou}
\alias{extract_ou}
\title{Read in OU Name from FAST PPL tab}
\usage{
extract_ou(filepath, cell)
}
\arguments{
\item{filepath}{full file path for the COP19 FAST}

\item{cell}{cell to extract OU name from}
}
\description{
Read in OU Name from FAST PPL tab
}
