% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_TCGA_sample_type}
\alias{extract_TCGA_sample_type}
\title{extract_TCGA_sample_type}
\usage{
extract_TCGA_sample_type(barcodes, return = "long")
}
\description{
extract_TCGA_sample_type
}
\author{
Jiahao Wang
}
