% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{extractVDJ}
\alias{extractVDJ}
\title{extractVDJ}
\usage{
extractVDJ(vdjOutDir, type = NULL)

extractVDJ(vdjOutDir, type = NULL)
}
\description{
extractVDJ

extractVDJ
}
\author{
Jiahao Wang
}
