% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastr_fit.R
\name{fa_fit}
\alias{fa_fit}
\title{Fit a latent factor model to single or multiple neuron spike trains}
\usage{
fa_fit(
  data,
  dt,
  n_factor,
  init,
  method = "2step",
  lam = NULL,
  nu = 15,
  woodbury = T,
  silent = F,
  adfun_only = F,
  ...
)
}
\arguments{
\item{data}{A \verb{n_cell x n_bin x n_trial} (multiple neurons)  or \verb{n_bin x n_trial}
(single neuron) array of binary spike trains. See details for if the spike train
data is in the form of spike times.}

\item{dt}{Scalar. Length of each time bin for time discretization of spike trains.}

\item{n_factor}{Integer. Number of independent factors in factor analysis.}

\item{init}{A named list of initial values for the parameter vectors \code{log_a} and
\code{log_k}  both of length \code{n_cell}, and the lower triangular entries of the loading
matrix \code{Lt} of length \code{n_cell*n_factor-n_factor*(n_factor-1)/2}.
If \code{method="2steps"}, only the initial values for the lower triangular loading
matrix needs to be provided in the form of \code{list(Lt = Lt_init_values)}.}

\item{method}{Estimation method: "2step" or "joint".
Default is "2step" : the first step is estimating drift \code{a} and threshold \code{k}
marginally using an inverse Gaussian model for each neuron independently, and the
second step is estimating the factor loading matrix with \code{a_hat} and \code{k_hat} as
plug-ins the joint likelihood. "joint" method estimates all parameters jointly,
hence more computationally intensive.}

\item{lam}{Scalar. Penalization parameter for the loading matrix elements. Default
is \code{lam=ifelse(n_cell<10, 1, 0.5)}.}

\item{nu}{Scalar. Parameter that controls the "steepness" around 0 of the sigmoid
function applied at the spike data likelihood layer. Default is 15.}

\item{woodbury}{Use the Woodbury matrix identity to speed up the calculation of
the covariance matrix? Default is T.}

\item{silent}{Suppress model fitting messages?}

\item{adfun_only}{Only outputs of ADFun created by TMB? This is for debugging
purposes only.}

\item{...}{Additional arguments to be passed to the optimization function
\code{nlminb()}}
}
\value{
A list of class \code{fastr_fit} containing the following objects:
\itemize{
\item time: time taken for model fitting,
\item log_a_hat: estimates of log drift parameters \code{a},
\item log_k_hat: estimates of log threshold parameters \code{k},
\item loga_se: standard errors of \code{log_a_hat},
\item logk_se: standard errors of \code{log_k_hat},
\item lmat_hat: estimate of the loading matrix L,
\item lmat_basis_cov: covariance matrix of the raw computational basis of L,
\item lmat_varimax: Varimax-transformed loading matrix estimate,
\item env: Other objects in the environment created during model fitting.
}
}
\description{
Fit a latent factor model to single or multiple neuron spike trains
}
\details{
If the spike train data comes in the form of spike times, it can be
converted to the format accepted by \code{fa_fit()} using the function \code{num2bin()}.
}
