\name{fastR-package}
\alias{fastR-package}
\alias{fastR}
\docType{package}
\title{Foundations and Applications of Statistics}

\description{
Data sets and utility functions to accompany
\cite{Foundations and Applications of Statistics: An Introduction Using R}
by Randall Pruim.
}
\details{
\tabular{ll}{
Package: \tab fastR\cr
Type: \tab Package\cr
Version: \tab 0.1-30\cr
Date: \tab 2010-08-19\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
Randall Pruim

Maintainer: Randall Pruim <rpruim@calvin.edu>
}
\references{
R. Pruim, 
\cite{Foundations and Applicaitons of Statistics: An Introduction Using R},
AMS, 2011, to appear.
}
\keyword{ package, datasets }
\seealso{
Other packages used in the text include:
\code{\link[lattice:lattice-package]{lattice}},
\code{\link[grid:grid-package]{grid}},
\code{\link[faraway:faraway-package]{faraway}},
\code{\link[Hmisc:Hmisc-package]{Hmisc}},
\code{\link[DAAG:DAAG-package]{DAAG}},
\code{\link[MASS:MASS-package]{MASS}},
\code{\link[multcomp:multcomp-package]{multcomp}},
\code{\link[vcd:vcd-package]{vcd}},
\code{\link[corrgram:corrgram-package]{corrgram}},
\code{\link[BradleyTerry:BradleyTerry-package]{BradleyTerry}},
\code{\link[alr3:alr3-package]{alr3}},
and
\code{\link[car:car-package]{car}} 
}
\examples{
None yet.
}
