\name{fastr}
\alias{fastr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fastr
}
\description{
Functions for interacting with the FastR implementation
}
\usage{
fastr.compile(f)
}

\details{

}
\value{
invisble NULL
}
\references{

}
\author{
FastR Team
}
\note{

}

\seealso{
%
}
\examples{
fastr.compile(f)
}
