% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{FeaturePlot2}
\alias{FeaturePlot2}
\title{FeaturePlot2}
\usage{
FeaturePlot2(
  obj,
  reduction = "umap",
  features,
  axis = TRUE,
  miniAxis = FALSE,
  legend = TRUE
)

FeaturePlot2(
  obj,
  reduction = "umap",
  features,
  axis = TRUE,
  miniAxis = FALSE,
  legend = TRUE
)
}
\description{
FeaturePlot2

FeaturePlot2
}
\author{
Jiahao Wang
}
