% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_cor_mat.R
\name{fill_cor_mat}
\alias{fill_cor_mat}
\title{Fill a correlation matrix with a vector of its lower triangular elements}
\usage{
fill_cor_mat(lower_tri, dim)
}
\arguments{
\item{lower_tri}{Vector of lower triangular elements (by column) of the correlation matrix}

\item{dim}{Number of rows/columns of the correlation matrix}
}
\value{
A dim x dim correlation matrix
}
\description{
Fill a correlation matrix with a vector of its lower triangular elements
}
