% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FindAllMarkers_genesort}
\alias{FindAllMarkers_genesort}
\title{FindAllMarkers_genesort}
\usage{
FindAllMarkers_genesort(obj, ntop = 100, group.by = NULL)
}
\description{
FindAllMarkers_genesort
}
\author{
Jiahao Wang
}
