% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_codb.R
\name{gather_codb}
\alias{gather_codb}
\title{Reshape budget columns long for CODB tab}
\usage{
gather_codb(df)
}
\arguments{
\item{df}{FAST data frame to gather}
}
\description{
Reshape budget columns long for CODB tab
}
