% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_cross.R
\name{gather_cross}
\alias{gather_cross}
\title{Reshape budget columns long for Cross Cutting tab}
\usage{
gather_cross(df, filepath)
}
\arguments{
\item{df}{FAST data frame to gather}

\item{filepath}{full file path for the COP19 FAST}
}
\description{
Reshape budget columns long for Cross Cutting tab
}
