% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_init.R
\name{gather_init}
\alias{gather_init}
\title{Reshape budget columns long for Initiatives tab}
\usage{
gather_init(df)
}
\arguments{
\item{df}{FAST data frame to gather}
}
\description{
Reshape budget columns long for Initiatives tab
}
