% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_inter_cop20.R
\name{gather_inter_cop20}
\alias{gather_inter_cop20}
\title{Reshape budget columns long}
\usage{
gather_inter_cop20(df)
}
\arguments{
\item{df}{FAST data frame to gather}
}
\description{
Reshape budget columns long
}
