% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FA_estim.R
\name{get_FA_estim}
\alias{get_FA_estim}
\title{Obtain L and a/k from model fit}
\usage{
get_FA_estim(mod_fit, n_cell, n_factor)
}
\arguments{
\item{mod_fit}{Fitted factor model output by nlminb(adfun)}

\item{n_cell}{Number of neurons}

\item{n_factor}{Number of factors}
}
\value{
A list of normalized L matrix and the firing rate ratio (a/k)
}
\description{
Obtain L and a/k from model fit
}
