% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_GEO_group}
\alias{get_GEO_group}
\title{get_GEO_group}
\usage{
get_GEO_group(
  obj,
  group_name,
  case_text,
  case_name,
  control_text,
  control_name
)
}
\description{
get_GEO_group
}
\author{
Jiahao Wang
}
