% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ig_mle.R
\name{get_ig_mle}
\alias{get_ig_mle}
\title{Calculate the MLEs for mu and lambda of an inverse Gaussian distribution}
\usage{
get_ig_mle(Y, dt)
}
\arguments{
\item{Y}{A \verb{q x n x r} or \verb{n x r} array of 0/1 spike trains where \code{q} is the number of cells,}

\item{dt}{Scalar of length of the bins.}
}
\value{
A list of MLEs of mu and lambda.
}
\description{
Calculate the MLEs for mu and lambda of an inverse Gaussian distribution
}
\details{
Let T be the vector of all ISIs. MLEs are calculated by\preformatted{\\hat\{\\mu\} = \\bar\{T\}, \\hat\{\\lambda\} = n/\\sum_\{i=1\}^n(1/T_i - 1/\\bar\{T\}).
}
}
