% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_versions}
\alias{get_package_versions}
\title{get_package_versions}
\usage{
get_package_versions(..., output_file = "package_versions.csv")
}
\description{
get_package_versions
}
\author{
Jiahao Wang
}
