% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spk_times_from_path.R
\name{get_spk_times_from_path}
\alias{get_spk_times_from_path}
\title{Get spike times from the latent path and threshold.}
\usage{
get_spk_times_from_path(path, k)
}
\arguments{
\item{path}{Length \code{N} vector of the path.}

\item{k}{Scalar threshold value for spike generating when the path crosses
the threshold.}
}
\value{
A vector of integers representing at which bins spikes occur.
}
\description{
Get spike times from the latent path and threshold.
}
