% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{getMultipletRate}
\alias{getMultipletRate}
\title{getMultipletRate}
\usage{
getMultipletRate(obj)

getMultipletRate(obj)
}
\description{
getMultipletRate

getMultipletRate
}
\author{
Jiahao Wang
}
