% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_style}
\alias{gg_style}
\title{gg_style}
\usage{
gg_style(
  title.size = 10,
  title.face = "bold",
  title.hjust = 0.5,
  title.vjust = 0,
  x.title.size = NA,
  x.title.face = NA,
  x.title.vjust = 0.5,
  x.text.size = NA,
  x.text.angle = 0,
  x.text.hjust = 0.5,
  x.text.vjust = 1,
  y.title.size = NA,
  y.title.vjust = 2,
  y.text.size = NA,
  face = "plain",
  y.title.face = NA,
  y.text.angle = 0,
  y.text.hjust = 1,
  legend.position = "right",
  plot.margin = c(5.5, 5.5, 5.5, 15),
  color = NULL,
  fill = NULL,
  theme = "classic"
)
}
\description{
gg_style
}
\author{
Jiahao Wang
}
