% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missing_inter.R
\name{handle_missing_inter}
\alias{handle_missing_inter}
\title{Handle missing interventions where there is a Total}
\usage{
handle_missing_inter(df)
}
\arguments{
\item{df}{}
}
\description{
Handle missing interventions where there is a Total
}
