% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_ou_cop20.R
\name{identify_ou_cop20}
\alias{identify_ou_cop20}
\title{Add Operating Unit into FAST Data Frame}
\usage{
identify_ou_cop20(df, filepath)
}
\arguments{
\item{df}{FAST data frame}

\item{filepath}{full file path for the COP19 FAST}
}
\description{
Add Operating Unit into FAST Data Frame
}
