% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fast.R
\name{import_fast}
\alias{import_fast}
\title{Import COP19 FAST Tool}
\usage{
import_fast(filepath, sheetname, skiprows = 2)
}
\arguments{
\item{filepath}{full file path for the COP19 FAST}

\item{sheetname}{name of the sheet/tab to import}

\item{skiprows}{how many rows to skip}
}
\description{
Import COP19 FAST Tool
}
