% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_headers.R
\docType{data}
\name{init_headers}
\alias{init_headers}
\title{Column Headers for the FAST}
\format{A vector of 68}
\usage{
init_headers
}
\description{
A vector of column headers to apply to the "3 Initiative-E" sheet of the FAST
}
\keyword{datasets}
