% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit function}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
Inverse logit of x, a number between 0 and 1.
}
\description{
Inverse logit function
}
