% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_oldformat.R
\name{is_oldformat}
\alias{is_oldformat}
\title{Check if the FAST is the old, preSGAC validated version}
\usage{
is_oldformat(filepath)
}
\arguments{
\item{filepath}{full file path for the COP19 FAST}
}
\description{
Check if the FAST is the old, preSGAC validated version
}
