% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_codb.R
\name{limit_codb}
\alias{limit_codb}
\title{Reduce to relevant rows and columns in CODB Tab}
\usage{
limit_codb(df)
}
\arguments{
\item{df}{FAST data frame to narrow down}
}
\description{
Reduce to relevant rows and columns in CODB Tab
}
