% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_comm.R
\name{limit_comm}
\alias{limit_comm}
\title{Reduce to relevant rows/columns and rename in Commodities Tab}
\usage{
limit_comm(df)
}
\arguments{
\item{df}{FAST data frame to narrow down}
}
\description{
Reduce to relevant rows/columns and rename in Commodities Tab
}
