% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_cross.R
\name{limit_cross}
\alias{limit_cross}
\title{Reduce to relevant rows and columns in Cross Cutting Tab}
\usage{
limit_cross(df, filepath)
}
\arguments{
\item{df}{FAST data frame to narrow down}

\item{filepath}{full file path for the COP19 FAST}
}
\description{
Reduce to relevant rows and columns in Cross Cutting Tab
}
