% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_fast.R
\name{limit_fast}
\alias{limit_fast}
\title{Reduce to relevant rows and columns}
\usage{
limit_fast(df)
}
\arguments{
\item{df}{FAST data frame to narrow down}
}
\description{
Reduce to relevant rows and columns
}
