% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_inter_cop20.R
\name{limit_inter_cop20}
\alias{limit_inter_cop20}
\title{Reduce to relevant rows and columns in Intervention Tab}
\usage{
limit_inter_cop20(df)
}
\arguments{
\item{df}{FAST data frame to narrow down}
}
\description{
Reduce to relevant rows and columns in Intervention Tab
}
