% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnfa-package.R
\docType{package}
\name{mnfa-package}
\alias{mnfa}
\alias{mnfa-package}
\title{mnfa: Multi-Neuron Factor Analysis}
\description{
This package provides an efficient implementation of a latent factor model for analyzing spike trains recorded from multiple neurons. Quantities of interest in this model are those that charaterize individual neuron's spiking intensity and those that represent the correlation among simultaneously recorded neurons.
}
\author{
\strong{Maintainer}: Meixi Chen \email{meixi.chen@uwaterloo.ca}

Authors:
\itemize{
  \item Martin Lysy
}

}
\keyword{internal}
