% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{multiPara}
\alias{multiPara}
\title{multiPara}
\usage{
multiPara(func, ..., paraL = NULL, bind = "c", cores = NULL)

multiPara(func, ..., paraL = NULL, bind = "c", cores = NULL)
}
\description{
multiPara

multiPara
}
\author{
Jiahao Wang
}
