% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_mat.R
\name{norm_mat}
\alias{norm_mat}
\title{Normalize the rows of L and Psi}
\usage{
norm_mat(L, psi, prec)
}
\arguments{
\item{L}{Loading matrix}

\item{psi}{Vector of diagonal elements of the error term matrix}

\item{prec}{Precision of the output}
}
\value{
A list containing the normalized L matrix and the diagonal vector of Psi, such that the norm of each row of L plus the diagonal of Psi equals 1
}
\description{
Normalize the rows of L and Psi
}
