% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pairwise_corr}
\alias{pairwise_corr}
\title{Pairwise correlation based on spike counts}
\usage{
pairwise_corr(neuron1, neuron2, block_len)
}
\arguments{
\item{neuron1}{A length \code{n} vector of neuron 1's binary spiking data in one trial.}

\item{neuron2}{A length \code{n} vector of neuron 2's binary spiking data in one trial.}

\item{block_len}{Number of bins to combine into blocks in which the number of spikes is counted.}
}
\value{
Correlation between two neurons' spiking counts
}
\description{
Pairwise correlation based on spike counts
}
