\name{petstress}
\alias{petstress}
\docType{data}
\title{
Pets and stress -- unfinished documentation
}
\description{
}
\usage{data(petstress)}
\format{
  A data frame with 45 observations on the following 2 variables.
  \describe{
    \item{\code{Group}}{a factor with levels \code{C} \code{F} \code{P}}
    \item{\code{Rate}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
data(petstress)
}
\keyword{datasets}
