% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pheatmap_DEG}
\alias{pheatmap_DEG}
\title{pheatmap_DEG}
\usage{
pheatmap_DEG(
  expM,
  DEG_tb,
  group,
  ntop = 20,
  out_name = "Heatmap_DEG",
  w = 6,
  h = 6,
  cluster_cols = FALSE,
  cluster_rows = FALSE,
  show_rownames = TRUE,
  show_colnames = FALSE
)
}
\description{
pheatmap_DEG
}
\author{
Jiahao Wang
}
