% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_coefficient_barplot}
\alias{plot_coefficient_barplot}
\title{plot_coefficient_barplot}
\usage{
plot_coefficient_barplot(
  multi_res,
  pvalue_cut = 0.05,
  out_name = "Coefficient_barplot"
)
}
\description{
plot_coefficient_barplot
}
\author{
Jiahao Wang
}
