% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_enrich_circle_KEGG}
\alias{plot_enrich_circle_KEGG}
\title{plot_enrich_circle_KEGG}
\usage{
plot_enrich_circle_KEGG(
  enrich_tb,
  gene_anno,
  terms = 10,
  n_gene = 20,
  space = 0.02,
  gene.space = 0.2,
  gene.size = 4,
  gene.col = c("firebrick3", "white", "steelblue")
)
}
\description{
plot_enrich_circle_KEGG
}
\author{
Jiahao Wang
}
