% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gof_spk_csum}
\alias{plot_gof_spk_csum}
\title{Observed versus predicted cumulative spike counts as as graphical GOF test}
\usage{
plot_gof_spk_csum(
  obs,
  pred,
  add_average = TRUE,
  xlab = "Observed cumulative spike count",
  ylab = "Predicted cumulative spike count",
  title = "",
  alpha_line = 0.5,
  ...
)
}
\arguments{
\item{obs}{Length \code{n_bin} vector of observed spikes, where each element is
a binary indicator of whether a spike has occurred in that bin.}

\item{pred}{\verb{n_bin x n_sample} matrix of predicted spikes. Each column
corresponds to a sample from the leave-neuron-out posterior predictive
distribution, obtained using \code{\link[=leave_one_out_prediction]{leave_one_out_prediction()}}.}

\item{add_average}{Add the average of all lines?}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{title}{Optional.}

\item{alpha_line}{Transparency for color of the lines.}

\item{...}{Additional arguments to the \code{plot()} function.}
}
\value{
Plot whose the X-axis corresponds to the observed
cumulative spike count and Y-axis corresponds to the predicted cumulative
spike counts. The counts are joined by lines, with each dot plotting the
predicted cumulative spike count from the LNO-PP distribution against
the observed value.
If \code{add_average=TRUE}, a red line corresponding to average of all lines is
added to the plot.
}
\description{
Observed versus predicted cumulative spike counts as as graphical GOF test
}
