% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gof_spk_times}
\alias{plot_gof_spk_times}
\title{Observed spike time vs histogram of predicted spike times as a graphical GOF test}
\usage{
plot_gof_spk_times(
  obs,
  pred,
  dt,
  spk_ind = NULL,
  plot_type = c("density", "hist", "none"),
  grid_plot = FALSE,
  pvals = FALSE
)
}
\arguments{
\item{obs}{Length \code{n_bin} vector of observed spikes, where each element is
a binary indicator of whether a spike has occurred in that bin.}

\item{pred}{\verb{n_bin x n_sample} matrix of predicted spikes. Each column
corresponds to a sample from the leave-neuron-out posterior predictive
distribution, obtained using \code{\link[=leave_one_out_prediction]{leave_one_out_prediction()}}.
Alternatively, a length \code{n_sample} list of predicted spike times can be
provided.}

\item{dt}{Scalar time scale of the spike trains. I.e., what is the length of
each bin in secs.
If spike times are provided for \code{pred}, make sure they are the same time scale
as \code{dt}.}

\item{spk_ind}{A vector of integers \eqn{I_1,\ldots,I_N}. The spike times for
each of the \eqn{I_n}-th spike are plotted. If not supplied, 4 spikes are
randomly selected.}

\item{plot_type}{"density" or "hist"?}

\item{grid_plot}{Plot spike time histogram/density plots in different grid
cells for different spike indices? Default is FALSE.}

\item{pvals}{Output p-values P(pred > obs)? Default is FALSE}
}
\value{
A ggplot object of histogram or density plot.
}
\description{
Observed spike time vs histogram of predicted spike times as a graphical GOF test
}
