% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_path}
\alias{plot_path}
\title{Plot the resetting latent path of a neuron with the 95\% CI}
\usage{
plot_path(
  path,
  se,
  reset = FALSE,
  spks = NULL,
  k = NULL,
  line_size = 1.2,
  alpha = 0.1,
  ci_fill = "green",
  ci_col = "black",
  pt_col = "red",
  pt_size = 1.2,
  hline_col = "blue",
  ann_coor = c(1, k + 0.1),
  ann_size = 8,
  ...
)
}
\arguments{
\item{path}{A vector of the latent path of neuron. This can be either the non-reset path
or the reset path (specify using the \code{reset} argument).}

\item{se}{A \code{length(path)} vector of the point-wise SE of the latent path.}

\item{reset}{TRUE if \code{path} is the reset path, i.e. not resetting needs to be performed.}

\item{spks}{An optional \code{length(path)} vector of 0s and 1s indicating spike occurrences.
Must be provided if \code{reset=FALSE}.}

\item{k}{Value of the threshold of excitation. Optional. If provided, a horizontal line will
be drawn indicating k.}

\item{line_size}{Size of the path in the plot.}

\item{alpha}{Transparency of the shaded credible intervals.}

\item{ci_fill}{Fill color for the shaded credible intervals.}

\item{ci_col}{Line color for the credible intervals.}

\item{pt_col}{Color of the points indicating spike occurrences.}

\item{pt_size}{Size of the points indicating spike occurrences.}

\item{hline_col}{Color of the horizontal line indicating k.}

\item{ann_coor}{A length 2 vector of the coordinate of the annotation "k".}

\item{ann_size}{Size of the annotation of "k".}
}
\value{
A ggplot plot.
}
\description{
Plot the resetting latent path of a neuron with the 95\% CI
}
