% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_survival_forest}
\alias{plot_survival_forest}
\title{plot_survival_forest}
\usage{
plot_survival_forest(
  pdata,
  out_name = "Survival_forest",
  cols = c("#F38329", "#1C79B7"),
  h = NULL,
  top = 30
)
}
\description{
plot_survival_forest
}
\author{
Jiahao Wang
}
