% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_volcano_limma}
\alias{plot_volcano_limma}
\title{plot_volcano_limma}
\usage{
plot_volcano_limma(
  DEG_tb,
  log2FC = 1,
  pvalue = 0.05,
  title = "",
  xtitle = "log2(Fold Change)",
  alpha = 0.5,
  padj = TRUE,
  text.cex = 20,
  label.cex = 2,
  pt.size = 0.8,
  label = TRUE,
  label.top = 10,
  w = 8,
  h = 7,
  out_name = NULL,
  max.overlaps = Inf,
  method = "pvalue",
  caption = FALSE,
  xlims = NULL,
  ylims = NULL
)
}
\description{
plot_volcano_limma
}
\author{
Jiahao Wang
}
