% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{processSigMarker}
\alias{processSigMarker}
\title{processSigMarker}
\usage{
processSigMarker(object, i, n = 50, write = F)

processSigMarker(object, i, n = 50, write = F)
}
\description{
processSigMarker

processSigMarker
}
\author{
Jiahao Wang
}
