% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_paths.R
\name{prop_paths}
\alias{prop_paths}
\title{Propose an initial path (x) given parameters k and alpha}
\usage{
prop_paths(Y, dt, log_k, log_a)
}
\arguments{
\item{Y}{A n x r array or q x n x r array of data where q is num of neurons, n is num of bins, r is num of trials}

\item{dt}{Length of time bin}

\item{log_k}{Length q vector of log threshold parameters}

\item{log_a}{Length q vector of log drift parameters}
}
\description{
Propose an initial path (x) given parameters k and alpha
}
