% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterplot.R
\name{rasterplot}
\alias{rasterplot}
\title{Raster plot of spike trains}
\usage{
rasterplot(data, ...)
}
\arguments{
\item{data}{A \code{r} by \code{n} matrix / data frame, or a list of \code{r} vectors.
If a matrix / df is provided, each row is a trial discretized into \code{n}
time bins in which a spike is indicated by 1 and 0 otherwise.
If a list is provided, each vector is a trial and contains the spike times.}

\item{...}{Additional arguments to pass to \code{plot()}.}
}
\value{
A raster plot
}
\description{
Raster plot of spike trains
}
