% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{readfile2seurat}
\alias{readfile2seurat}
\title{readfile2seurat}
\usage{
readfile2seurat(file_path)

readfile2seurat(file_path)
}
\description{
readfile2seurat

readfile2seurat
}
\author{
Jiahao Wang
}
