% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{readRData}
\alias{readRData}
\title{readRData}
\usage{
readRData(file_path, idx = NULL, rm = FALSE, verbose = FALSE)

readRData(file_path, idx = NULL, rm = FALSE, verbose = FALSE)
}
\description{
readRData

readRData
}
\author{
Jiahao Wang
}
