% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{real2cor}
\alias{real2cor}
\title{Transform a real number to a correlation coefficient}
\usage{
real2cor(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
A number between -1 and 1.
}
\description{
Transform a real number to a correlation coefficient
}
\details{
This function is the inverse function of \code{cor2real()}.
}
