% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/utils.R
\name{report}
\alias{report}
\title{report}
\usage{
report(info, type = "I", print = TRUE, file = NULL, append = TRUE)

report(info, type = "I", print = TRUE, file = NULL, append = TRUE)
}
\description{
report

report
}
\author{
Jiahao Wang
}
