% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_path.R
\name{reset_path}
\alias{reset_path}
\title{Reset the latent path}
\usage{
reset_path(path, spks)
}
\arguments{
\item{path}{A vector of the path which has not been reset.}

\item{spks}{A vector of 0s and 1s for spikes, which must be the same
length as \code{path}.}
}
\value{
A vector of reset path according to the spikes.
}
\description{
Reset the latent path
}
