% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rmvn_prec}
\alias{rmvn_prec}
\title{Sample from a multivariate normal with sparse precision matrix.}
\usage{
rmvn_prec(n, mean, prec)
}
\arguments{
\item{n}{Number of random draws.}

\item{mean}{Mean vector.}

\item{prec}{Sparse precision matrix, i.e., inheriting from
\link[Matrix:sparseMatrix]{Matrix::sparseMatrix} or its Cholesky factor, i.e.,
inheriting from "CHMfactor".}
}
\value{
A matrix with \code{n} rows, each of which is a draw from the
corresponding normal distribution.
}
\description{
Sample from a multivariate normal with sparse precision matrix.
}
\details{
If the matrix is provided in precision form, it is converted to
Cholesky form using \code{Matrix::Cholesky(prec, super = TRUE)}.  Once it is of
form "CHMfactor", this function is essentially copied from local
function \code{rmvnorm()} in function \code{MC()} defined in \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.
}
