% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_check_point}
\alias{run_check_point}
\title{run_check_point}
\usage{
run_check_point(
  expM,
  group = NULL,
  group_name = "Group",
  gene_list = NULL,
  prefix = "./Check_point_",
  Rplot = FALSE
)
}
\description{
run_check_point
}
\author{
Jiahao Wang
}
