% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_core_pathway}
\alias{run_core_pathway}
\title{run_core_pathway}
\usage{
run_core_pathway(
  expM,
  group = NULL,
  group_name = "Group",
  gene_list,
  prefix = "./Core_pathway",
  Rplot = FALSE
)
}
\description{
run_core_pathway
}
\author{
Jiahao Wang
}
