% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_DEG_deseq2}
\alias{run_DEG_deseq2}
\title{run_DEG_deseq2}
\usage{
run_DEG_deseq2(
  expM,
  Group,
  Case,
  Control,
  out_name = NULL,
  log2FC = 1,
  pvalue = 0.05,
  padj = TRUE
)
}
\description{
run_DEG_deseq2
}
\author{
Jiahao Wang
}
