% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_DEG_limma}
\alias{run_DEG_limma}
\title{run_DEG_limma}
\usage{
run_DEG_limma(
  expM,
  Group,
  Case,
  Control,
  out_name = "DEG",
  log2FC = 1,
  pvalue = 0.05,
  padj = TRUE
)
}
\description{
run_DEG_limma
}
\author{
Jiahao Wang
}
