% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_estimate}
\alias{run_estimate}
\title{run_estimate}
\usage{
run_estimate(
  expM,
  group = NULL,
  group_name = "Group",
  prefix = "./ESTIMATE_",
  Rplot = FALSE
)
}
\description{
run_estimate
}
\author{
Jiahao Wang
}
