% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fast.R
\name{run_fast}
\alias{run_fast}
\title{Run the COP19 FAST Importer}
\usage{
run_fast(filepath, ou = NULL)
}
\arguments{
\item{filepath}{full file path for the COP19 FAST}

\item{ou}{specify the operatingunit}
}
\description{
Run the COP19 FAST Importer
}
\examples{
\dontrun{
#FAST file path
  path <- "../Downloads/FAST Malawi Consolidated Zero Draft 02082019.xlsx"
#read in FAST
  df_dp <- tame_dp(path, "Malawi") }
}
