% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_immune_infiltration}
\alias{run_immune_infiltration}
\title{run_immune_infiltration}
\usage{
run_immune_infiltration(
  expM,
  group = NULL,
  group_name = "Group",
  gene_list,
  prefix = "./Immune_infiltration",
  Rplot = FALSE
)
}
\description{
run_immune_infiltration
}
\author{
Jiahao Wang
}
