% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_ML_Boruta}
\alias{run_ML_Boruta}
\title{run_ML_Boruta}
\usage{
run_ML_Boruta(
  expM,
  Group,
  out_name = "Boruta",
  lambda = "min",
  seed = 1234,
  maxRuns = 100,
  pValue = 0.01
)
}
\description{
run_ML_Boruta
}
\author{
Jiahao Wang
}
