% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_ML_Lasso}
\alias{run_ML_Lasso}
\title{run_ML_Lasso}
\usage{
run_ML_Lasso(
  expM,
  Group,
  out_name = "Lasso",
  lambda = "min",
  seed = 1234,
  nfolds = 10,
  legend = TRUE
)
}
\description{
run_ML_Lasso
}
\author{
Jiahao Wang
}
