% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_multicox_survival}
\alias{run_multicox_survival}
\title{run_multicox_survival}
\usage{
run_multicox_survival(tdata, genes, out_name = "multiCox", pvalue_cut = 0.05)
}
\description{
run_multicox_survival
}
\author{
Jiahao Wang
}
