% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_PCA}
\alias{run_PCA}
\title{run_PCA}
\usage{
run_PCA(
  expM,
  Group,
  out_name = NULL,
  text.size = 20,
  pt.size = 1,
  ellipse = TRUE,
  w = 5,
  h = 5,
  legend.ncol = NULL,
  color = c("firebrick3", "steelblue")
)
}
\description{
run_PCA
}
\author{
Jiahao Wang
}
