% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_survival_lasso}
\alias{run_survival_lasso}
\title{run_survival_lasso}
\usage{
run_survival_lasso(tdata, genes, out_name = "Survial_LASSO", seed = 1234)
}
\description{
run_survival_lasso
}
\author{
Jiahao Wang
}
