% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_unicox_survival}
\alias{run_unicox_survival}
\title{run_unicox_survival}
\usage{
run_unicox_survival(tdata, genes, out_name, pvalue_cut = 0.05)
}
\description{
run_unicox_survival
}
\author{
Jiahao Wang
}
