% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runDO}
\alias{runDO}
\title{runDO}
\usage{
runDO(
  sigG,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  universe = NULL,
  out = NULL,
  ont = "DO",
  sig = TRUE,
  species = "human",
  raw = FALSE
)
}
\description{
runDO
}
\author{
Jiahao Wang
}
