% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R, R/utils.R
\name{runEGO}
\alias{runEGO}
\title{runEGO}
\usage{
runEGO(
  sigG,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  universe = NULL,
  out = NULL,
  keyType = "ENSEMBL",
  ont = "ALL",
  sig = TRUE,
  species = "human",
  raw = FALSE
)

runEGO(
  sigG,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  universe = NULL,
  out = NULL,
  keyType = "ENSEMBL",
  ont = "ALL",
  sig = TRUE,
  species = "human",
  raw = FALSE
)
}
\description{
runEGO

runEGO
}
\author{
Jiahao Wang
}
