% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R, R/utils.R
\name{runKEGG}
\alias{runKEGG}
\title{runKEGG}
\usage{
runKEGG(
  sigG,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  species = "human",
  out = NULL
)

runKEGG(
  sigG,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  species = "human",
  out = NULL
)
}
\description{
runKEGG

runKEGG
}
\author{
Jiahao Wang
}
