% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/visualization.R
\name{save.graph}
\alias{save.graph}
\title{save.graph}
\usage{
save.graph(p, file, w = 4, h = 4, mar = NULL, PDF = TRUE, PNG = TRUE)

save.graph(p, file, w = 4, h = 4, mar = NULL, PDF = TRUE, PNG = TRUE)
}
\description{
save.graph

save.graph
}
\author{
Jiahao Wang
}
