% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{sc_dotplot_general}
\alias{sc_dotplot_general}
\title{sc_dotplot_general}
\usage{
sc_dotplot_general(
  obj,
  group,
  file,
  PDF = TRUE,
  PNG = TRUE,
  species = "human",
  w = 10,
  h = 9
)

sc_dotplot_general(
  obj,
  group,
  file,
  PDF = TRUE,
  PNG = TRUE,
  species = "human",
  w = 10,
  h = 9
)
}
\description{
sc_dotplot_general

sc_dotplot_general
}
\author{
Jiahao Wang
}
