% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{sc_merge_qc}
\alias{sc_merge_qc}
\title{sc_merge_qc}
\usage{
sc_merge_qc(input_dir, type = "count")

sc_merge_qc(input_dir, type = "count")
}
\description{
sc_merge_qc

sc_merge_qc
}
\author{
Jiahao Wang
}
