% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sc_mini_axis}
\alias{sc_mini_axis}
\title{sc_mini_axis}
\usage{
sc_mini_axis(
  p,
  obj,
  reduction = "umap",
  pos.just = 1,
  label.just = 0.93,
  axis.len = 0.2,
  axis.label.cex = 3,
  arrow.len = 0.3,
  arrow.size = 0.5
)
}
\description{
sc_mini_axis
}
\author{
Jiahao Wang
}
