% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sc_pipeline}
\alias{sc_pipeline}
\title{sc_pipeline}
\usage{
sc_pipeline(
  obj_raw,
  npc = 30,
  ncluster = c(14, 16),
  resolution = 0.2,
  sample_name = "SID",
  out_name = "sc_pipeline",
  cell.min = 3,
  mt.pct = 10,
  hb.pct = 3,
  feature.min = 200,
  feature.max = 8000,
  tSNE = TRUE,
  count.min = 0,
  count.max = Inf,
  species = "human",
  verbose = FALSE
)
}
\description{
sc_pipeline
}
\author{
Jiahao Wang
}
