% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_seq.R, R/utils.R
\name{selectMarkers}
\alias{selectMarkers}
\title{selectMarkers}
\usage{
selectMarkers(celltype = NULL)

selectMarkers(celltype = NULL)
}
\description{
selectMarkers

selectMarkers
}
\author{
Jiahao Wang
}
