% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seurat2cds}
\alias{seurat2cds}
\title{seurat2cds}
\usage{
seurat2cds(
  obj,
  group.by,
  out_dir,
  ordering_genes = NULL,
  method = "marker",
  add.genes = NULL
)
}
\description{
seurat2cds
}
\author{
Jiahao Wang
}
