% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigm.R
\name{sigm}
\alias{sigm}
\title{Sigmoid function with mean and scale parameters}
\usage{
sigm(mu, sig, x)
}
\arguments{
\item{mu}{Mean parameter.}

\item{sig}{Scale parameter.}

\item{x}{Input to the sigmoid function, a real number.}
}
\value{
Sigmoid transformation of x, a number between 0 and 1.
}
\description{
Sigmoid function with mean and scale parameters
}
