% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Simulate spike trains given parameters}
\usage{
simdata(dt, n_bin, n_trial, alpha, k, rho, L)
}
\arguments{
\item{dt}{Time bin length.}

\item{n_bin}{Number of time bins.}

\item{n_trial}{Number of trials.}

\item{alpha}{Vector of drift parameters.}

\item{k}{Vector of threshold parameters.}

\item{rho}{Vector of correlation parameters of length q(q-1)/2 where q is the length of alpha.
If this is provided, the pairwise correlation model is used for simulation.}

\item{L}{Loading matrix of dimension q x d, where d is the number of latent factors.
If this is provided, the factor model is used.}
}
\description{
Simulate spike trains given parameters
}
