% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_output.R
\name{simplify_output}
\alias{simplify_output}
\title{Makes simplified output from an object of class \code{fastr_fit}}
\usage{
simplify_output(fobj, ...)
}
\arguments{
\item{fobj}{An object of class \code{fastr_fit}.}

\item{...}{Additional arguments.}
}
\value{
A list of class \code{fastr_fit}.
}
\description{
Makes simplified output from an object of class \code{fastr_fit}
}
\details{
This is both a helper function used within the function \code{fastr_fit()} and
can also be used standalone.
}
