% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R, R/utils.R
\name{subString}
\alias{subString}
\title{subString}
\usage{
subString(strings, idx, sep = NULL, rev = FALSE, collapse = NULL)

subString(strings, idx, sep = NULL, rev = FALSE, collapse = NULL)
}
\description{
subString

subString
}
\author{
Jiahao Wang
}
