% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_n_factor.R
\name{choose_n_factor}
\alias{choose_n_factor}
\title{Perform exploratory factor analysis to choose the number of factors}
\usage{
choose_n_factor(
  data,
  dt,
  d_seq,
  fa_obj = NULL,
  plot = TRUE,
  vertical = TRUE,
  progress = TRUE,
  ...
)
}
\arguments{
\item{data}{A \verb{q x d x r} array, where \code{q} is the number of neurons, \code{d} is
the number of time bins, \code{r} is the number of trials.}

\item{dt}{Scalar. Length of each time bin for time discretization.}

\item{d_seq}{A vector of candidate \code{n_factor} values. See details.}

\item{fa_obj}{A \verb{q x dr} matrix of increments from the result of a previous
run of \code{choose_n_factor()}. This argument is helpful when we want to test a
different set of \code{d_seq} but don't need to fit all single-neuron models again.}

\item{plot}{Output a plot of cumulative var explained versus \code{d}?}

\item{vertical}{If `plot=TRUE", should the plots be arranged vertically (or
horizontally)?}

\item{progress}{Print the fitting progress?}

\item{...}{Additional arguments supplied to \code{fastr_fit()} for fitting single-neuron
models.}
}
\value{
\enumerate{
\item A vector of length \code{length(d_seq)} containing the cumulative proportion of
variance explained by the corresponding number of factors (d).
\item A plot that shows the relationship between the two.
\item An array or a vector of all the single-neuron model fitted paths.
}
}
\description{
Perform exploratory factor analysis to choose the number of factors
}
\details{
Given data with \code{q} neurons, a single-neuron model is fitted to each neuron
to obtain \code{q} estimates of neuronal latent paths x_i for i=1,..., q. Then a factor
analysis model with \code{d} factor is fitted to the matrix whose column i is the latent
path increments dx_ij for neuron i. For each candidate \code{d} value in \code{d_seq}, the
cumulative proportion of variance explained is extracted. The function (optionally)
outputs a plot whose y-axis is the cumulative var and x-zxis is \code{d}.
}
