\name{fastR-package}
\alias{fastR-package}
\docType{package}
\title{
Really fast R functions
}

\description{
A collection of fast functions for data analysis.
}

\details{
\tabular{ll}{
Package: \tab fastR\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2016-06-27\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris <mtsagris@yahoo.gr> and Manos Papadakis <papadakm95@gmail.com>
}

\section{Maintainers }{
Michail Tsagris <mtsagris@yahoo.gr>
}


\keyword{ Mahalanobis distance }
\keyword{ covariance matrix }
\keyword{ cariances }
\keyword{ t-tests }
\keyword{ regression coefficients }


