% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastr-package.R
\docType{package}
\name{fastr-package}
\alias{fastr}
\alias{fastr-package}
\title{fastr: fastr: Factor analysis of multi-neuronal spike trains in R}
\description{
This package provides an efficient implementation of a latent factor model for analyzing spike trains recorded from multiple neurons. Quantities of interest in this model are those that charaterize individual neuron's spiking intensity and those that represent the correlation among simultaneously recorded neurons.
}
\author{
\strong{Maintainer}: Meixi Chen \email{meixi.chen@uwaterloo.ca}

Authors:
\itemize{
  \item Martin Lysy \email{mlysy@uwaterloo.ca}
}

Other contributors:
\itemize{
  \item Reza Ramezan \email{rramezan@uwaterloo.ca} [contributor]
}

}
\keyword{internal}
