\name{fastr-package}
\alias{fastr-package}
\alias{fastr}
\docType{package}
\title{Functions for interacting with the FastR implementation}
\description{Functions for interacting with the FastR implementation}
\details{
\tabular{ll}{
Package: \tab fastr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-05\cr
License: \tab GPL-2\cr
}

}
\author{
The FastR Team

Maintainer: fastr@yahoogroups.com
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{

}
