% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ig_mle.R
\name{get_ig_mle}
\alias{get_ig_mle}
\title{Calculate the MLEs for parameters of an inverse Gaussian distribution}
\usage{
get_ig_mle(Y, dt)
}
\arguments{
\item{Y}{A \verb{q x n x r} or \verb{n x r} array of 0/1 spike trains where \code{q} is the number of cells,
\code{n} is the number of time bins, and \code{r} is the number of trials.}

\item{dt}{Scalar of length of the bins.}
}
\value{
A list of MLEs and Hessian matrix of the (log) parameters c(log_k, log_a).
}
\description{
Calculate the MLEs for parameters of an inverse Gaussian distribution
}
\details{
Let T be the vector of all ISIs. MLEs are calculated by\preformatted{\\hat\{\\mu\} = \\bar\{T\}, \\hat\{\\lambda\} = n/\\sum_\{i=1\}^n(1/T_i - 1/\\bar\{T\}).
}

If this IG distribution represents the distribution of first passage time of a
Wiener process with threshold k, drift alpha and variance sigma^2, then mu =
k/alpha, lambda=k^2/sigma^2.
}
