% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_ou.R
\name{identify_ou}
\alias{identify_ou}
\title{Add Operating Unit into FAST Data Frame}
\usage{
identify_ou(df, ou)
}
\arguments{
\item{df}{FAST data frame}

\item{ou}{operatingunit}
}
\description{
Add Operating Unit into FAST Data Frame
}
