% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_init.R
\name{limit_init}
\alias{limit_init}
\title{Reduce to relevant rows and columns in Initiatives Tab}
\usage{
limit_init(df)
}
\arguments{
\item{df}{FAST data frame to narrow down}
}
\description{
Reduce to relevant rows and columns in Initiatives Tab
}
