% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnfa-package.R
\docType{package}
\name{mnfa-package}
\alias{mnfa}
\alias{mnfa-package}
\title{mnfa: Multi-Neuron Factor Analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Meixi Chen \email{meixi.chen@uwaterloo.ca}

Authors:
\itemize{
  \item Martin Lysy
}

}
\keyword{internal}
