\name{petstress}
\alias{petstress}
\docType{data}
\title{
Pets and stress -- unfinished documentation
}
\description{
Does having a pet or a friend cause more stress?
}
\usage{data(petstress)}
\format{
  A data frame with 45 observations on the following 2 variables.
  \describe{
    \item{\code{Group}}{a factor with levels \code{C}ontrol, \code{F}riend, or \code{P}et}
    \item{\code{Rate}}{a numeric vector}
  }
}
\examples{
data(petstress)
}
\keyword{datasets}
